Beta 1 (Techniche 2025 release)
This is in effect the final version of Madcow Decease that will be played at Techniche 2025. This is a working version in theory but has many bugs in practice.
These need to be resolved on a can do basis. Avoid pushing directly on main hereafter and clean up or make your own working branch.
Note that the level 1 music file is missing from the repo due to filesize issues. To get it, ask Shreyash.